backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.30.6
httpx==0.27.2
PyJWT==2.9.0
pydantic==2.9.2
pydantic-settings==2.4.0
boto3==1.35.54
